backend/requirements.txt
# Core framework
fastapi>=0.124.1
uvicorn[standard]==0.34.0
websockets>=15.0.1
python-dotenv==1.0.1
pydantic>=2.11.1
pydantic-settings>=2.7.1

# Google ADK & AI
google-adk==1.26.0

# GCP Services
google-cloud-firestore==2.20.1
google-cloud-storage==2.19.0

# PDF Report Generation
reportlab==4.2.5
Pillow==11.1.0

# Web search (for standards lookup tool — duckduckgo-search was renamed to ddgs)
ddgs>=1.1.0

# Utilities
httpx==0.28.1
python-multipart==0.0.20
aiofiles==24.1.0
